Fix rate-limit isolation, portfolio Decimal math, and dedupe ownership checks
Separates rate-limit keys per endpoint (crossover/zone/backtest) so scans
no longer share budget, isolates healthcheck failures to avoid duplicate
DB-outage alerts, fixes portfolio P&L to include priceless holdings via
Decimal fields, tracks dropped instruments in zone scan responses,
consolidates watchlist/screen ownership checks into shared deps helpers,
and adds ema_21 plus screen deep-linking from the command palette.

+++ app/models/holding.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import ForeignKey, Numeric, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column
@@ -14,6 +15,6 @@ class Holding(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
     instrument_id: Mapped[int] = mapped_column(ForeignKey("instruments.id", ondelete="CASCADE"), nullable=False)
-    quantity: Mapped[float] = mapped_column(Numeric(18, 4), nullable=False)
-    avg_cost: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
+    quantity: Mapped[Decimal] = mapped_column(Numeric(18, 4), nullable=False)
+    avg_cost: Mapped[Decimal] = mapped_column(Numeric(12, 4), nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, server_default=func.now())
